requirements.txt
annotated-types==0.7.0
anthropic==0.29.0
anyio==4.4.0
asgiref==3.8.1
asyncio==3.4.3
blinker==1.8.2
cachetools==5.3.3
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
distro==1.9.0
dnspython==2.6.1
docx2txt==0.8
exceptiongroup==1.2.1
filelock==3.15.4
Flask==3.0.3
fsspec==2024.6.0
google-ai-generativelanguage==0.6.5
google-api-core==2.19.1
google-api-python-client==2.134.0
google-auth==2.30.0
google-auth-httplib2==0.2.0
google-generativeai==0.7.0
googleapis-common-protos==1.63.2
groq==0.9.0
grpcio==1.64.1
grpcio-status==1.62.2
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
huggingface-hub==0.23.4
idna==3.7
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.5.0
lxml==5.2.2
MarkupSafe==2.1.5
motor==3.5.0
nest-asyncio==1.6.0
packaging==24.1
pluggy==1.5.0
proto-plus==1.24.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.7.4
pydantic_core==2.18.4
pymongo==4.7.3
pyparsing==3.1.2
PyPDF2==3.0.1
pytest==8.2.2
python-docx==1.1.2
python-dotenv==1.0.1
PyYAML==6.0.1
requests==2.32.3
rsa==4.9
sniffio==1.3.1
tokenizers==0.19.1
tomli==2.0.1
tqdm==4.66.4
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.2
vercel==0.2.1
Werkzeug==3.0.3
